pyproject.toml
[project]
name = "graffity_report"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.4.2",
    "flask>=3.1.2",
    "python-decouple>=3.8.0",
]
